+++ src/sentry/api/serializers/models/group_stream.py
@@ -17,11 +17,11 @@ from sentry.api.serializers.models.group import (
     SeenStats,
     snuba_tsdb,
 )
+from sentry.api.serializers.models.plugin import is_plugin_deprecated
 from sentry.constants import StatsPeriod
 from sentry.issues.grouptype import GroupCategory
 from sentry.models.environment import Environment
 from sentry.models.group import Group
-from sentry.models.groupactions import get_actions, get_available_issue_plugins
 from sentry.models.groupinbox import get_inbox_details
 from sentry.models.groupowner import get_owner_details
 from sentry.snuba.dataset import Dataset
@@ -29,7 +29,53 @@ from sentry.tsdb.base import TSDBModel
 from sentry.utils import metrics
 from sentry.utils.cache import cache
 from sentry.utils.hashlib import hash_values
+from sentry.utils.safe import safe_execute
 from sentry.utils.snuba import resolve_column, resolve_conditions
+
+
+def get_actions(request: Request, group):
+    from sentry.plugins.base import plugins
+
+    project = group.project
+
+    action_list = []
+    for plugin in plugins.for_project(project, version=1):
+        if is_plugin_deprecated(plugin, project):
+            continue
+
+        results = safe_execute(plugin.actions, request, group, action_list, _with_transaction=False)
+
+        if not results:
+            continue
+
+        action_list = results
+
+    for plugin in plugins.for_project(project, version=2):
+        if is_plugin_deprecated(plugin, project):
+            continue
+        for action in (
+            safe_execute(plugin.get_actions, request, group, _with_transaction=False) or ()
+        ):
+            action_list.append(action)
+
+    return action_list
+
+
+def get_available_issue_plugins(request: Request, group):
+    from sentry.plugins.base import plugins
+    from sentry.plugins.bases.issue2 import IssueTrackingPlugin2
+
+    project = group.project
+
+    plugin_issues = []
+    for plugin in plugins.for_project(project, version=1):
+        if isinstance(plugin, IssueTrackingPlugin2):
+            if is_plugin_deprecated(plugin, project):
+                continue
+            plugin_issues = safe_execute(
+                plugin.plugin_issues, request, group, plugin_issues, _with_transaction=False
+            )
+    return plugin_issues
 
 
 @dataclass
